requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
pydantic
python-dotenv

streamlit
requests

ipykernel
matplotlib
seaborn